Migrate translation status builder to a lib
